Grab only Unseen emails
In order to prevent a scheduled command that runs while a previous run
is happening from grabbing the same emails, we mark the messages as read
as soon as we grab them. Then if they relay we delete them, but if it
doesn't relay we mark the message as unseen for another try later.